Layout update
Layout of header-info
Set fixed width to header to
Align its edges with the ones
header-info.
Changes locations of selectors so the ones of
every section are at the same place
Added black sides to the look of the page by
setting margin on the main container, set
background the background color to the body.
Set the background color of the image placeholder
to reach the desired look.