Styling of header info, header width
Increased header section width since when
I did the styling to the header info  left section
, its height increased and then when I aligned the
img placeholder in order to make it look good I
increased its width. The total width of the header
info section elements increased and to align the
content of the header with it I increased its
width.